requirements.txt
# Restaurant Satisfaction Predictor - Streamlit + SQLite
streamlit>=1.28.0
python-dotenv>=1.0.0

# Pin versions - model was saved with sklearn 1.5+ (has _RemainderColsList)
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.5.0,<1.7.0
joblib>=1.3.0

# Model dependencies (logistic_regression, random_forest only)
